fix for parser bug, treating some identifiers as keywords
If an identifier contained a keyword as first letters, the token was
incorrectly parsed as keyword. Eg "falsee" was parsed as literal bool,
then sole "e" leading to very unclear syntax errors. This change fixes
this problem by adding a look-ahead after keywords to ensure there are
no letters left that would otherwise form an identifier.