bugfix uninitialized item_filter members
Added constructors to enforce initialization of members
which do not initialize themselves (eg bool) and do not
have good defaults to set.

Refactored block_continuation: boolean + origin was bad design
because if the boolean was false the origin had invalid state.
Now the only member is an optional origin and its existence
implies the boolean state.