simplify filter state mediators inheritance tree
Removed bases for mediator, real and spirit filter implementations.
They could not forward-declare some things and whenever any implementation
was changed, the interface would too which denied the purpose of having
the interface separately defined. Incremental build was almost always
rebuilding all of mediator files, and with extra intermediate code for
interfaces there was even more to parse and compile.